Fixed database error
Using a new command every time will allow for non-duplicate data to be
entered also using the same cmd every time results in an error because
it says that data is already set. CommandText had line breaks in it from
previous formatting way around this is to use a StringBuilder to
construct the CommandText which would also improve readability.
Parameter runwayAccel needs to match the same parameter in CommandText
and the connection for cmd must be set when doing the ExecuteNonQuery
function.