Fixed loading to the database
Fix #1. Data was not properly inserted into the data from the file (i.e.
key and value pairs did not match). The brackLocation key needed the
breakLocation value

Fix #2. The list that was returned from the database has blank spots at
the beginning that are supposed to house the direction, move, and track
circuit, these are skipped because they are not important to the data
being used.

Fix #3 Converted the algorithm for getting data from the list from O(2n)
to O(n) by using an offset instead of removing unwanted items from the
list.